separate handler and remove toggle for now
